fix(ecommerce-stripe): honor inStock=false, hoist product search schema
listProducts?inStock=false silently ignored the filter. inStock != null
made hasVariantLevelFilter true, so the variant-narrowing path ran, but
the guard was `if (inStock)` — falsy on false — so the stock check never
applied and products came back regardless of stock while the response
implied they had been filtered. Now tests `!= null` and compares the
per-variant stock result against the requested value, so true and false
both filter correctly.

Also moves ProductSearchQuerySchema out of the controller into
domain/schemas/productSearch.schema.ts, matching catalogImport.schema.ts
and checkout.schema.ts. This was the only remaining inline controller
schema in the stack.
